pyproject.toml
[project]
name = "novarch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.2",
    "typer>=0.17.3",
]
